Take ul/li specifics out of generic components
